pyproject.toml
[tool.poetry]
name = "apistudent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
httpie = "^3.2.2"
flask-pydantic-spec = "^0.5.0"
tinydb = "^4.8.0"
pymongo = "^4.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
